[GEOS-9826] Remove un-necessary casts from code, add a QA rule to enforce it

+++ src/ows/src/main/java/org/geoserver/ows/kvp/ElevationParser.java
@@ -92,7 +92,7 @@ public class ElevationParser {
                                 final NumberRange left = (NumberRange) o1;
                                 if (o2Double) {
                                     // o2 date
-                                    return left.getMinValue().compareTo(((Double) o2));
+                                    return left.getMinValue().compareTo(o2);
                                 }
                                 // o2 daterange
                                 return left.getMinValue()
